pyproject.toml
[project]
name = "task-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.117.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
]
